refactor: split global into preGame and inGame routes
