+++ src/routes/inGame/validators.js
@@ -1,4 +1,4 @@
-const { emote, isAllowed } = require('../../generalFunctions.js');
+const { emote, isAllowed, formatId } = require('../../generalFunctions.js');
 
 // globals
 let current;
@@ -92,4 +92,37 @@ function call(game, message) {
   }
   return true;
 }
-module.exports = { end, check, allIn, raise, fold, call };
+
+function join(game, message) {
+  if (
+    game != undefined &&
+    game.players[formatId(message.author)] !== undefined
+  ) {
+    message.reply('You have already joined!');
+    return false;
+  }
+  return true;
+}
+
+function show(game, message) {
+  if (game == undefined) {
+    message.react(emote('fold'));
+    message.reply('There are no players at the table');
+    return false;
+  }
+  return true;
+}
+
+function exit(game, message) {
+  if (
+    game === undefined ||
+    game.players[formatId(message.author)] === undefined
+  ) {
+    message.react(emote('mistake'));
+    message.reply('You have not joined the game yet');
+    return false;
+  }
+  return true;
+}
+
+module.exports = { end, check, allIn, raise, fold, call, join, show, exit };
