+++ src/routes/router.js
@@ -1,15 +1,14 @@
 // Scripts
-const constants = require("../constants");
-const globalRoute = require("./global/index.js");
-const preGameRoute = require("./preGame/index.js");
-const inGameRoute = require("./inGame/index.js");
+const constants = require('../constants');
+const preGameRoute = require('./preGame/index.js');
+const inGameRoute = require('./inGame/index.js');
 
 function validateMessage(message, body, chat) {
   const messageAge = Math.floor(Date.now() / 1000) - message.timestamp;
   if (
     messageAge < constants.MESSAGE_MAX_AGE &&
     chat.isGroup &&
-    body[0] == "pok"
+    body[0] == 'pok'
   ) {
     return true;
   }
@@ -20,12 +19,10 @@ async function route(whatsapp, message, body, chat, games) {
   const chatId = chat.id.user;
   const phoneNumber = (await message.getContact()).id.user;
 
-  if (globalRoute(body, games, chatId, message, phoneNumber, chat)) {
-    return true;
-  } else if (games[chatId] == undefined || !games[chatId].isMidRound) {
-    preGameRoute(body, games[chatId], message, whatsapp);
+  if (games[chatId] == undefined || !games[chatId].isMidRound) {
+    preGameRoute(body, games, chatId, message, whatsapp, chat, phoneNumber);
   } else {
-    inGameRoute(body, games, chatId, message, whatsapp);
+    inGameRoute(body, games, chatId, message, whatsapp, chat, phoneNumber);
   }
 }
